 \docType{data}
    \name{taylor}
    \alias{taylor}
    \title{The MSKCC Cohort from Taylor et al. (2010)}
    \format{a \code{ExpressionSet} instance, 1 row per Affymetrix microaray probe.}
    \source{
      GSE21034
    }
    \description{
      Object imported using GEOquery and curated to remove variables introduced by GEO. Details in Vignette.
    }
        \value{ExpressionSet object for the dataset}
    \author{
Mark Dunning
    }
    \usage{camcap}
    \keyword{datasets}